copy settings on index clone

+++ src/clean_indices.py
@@ -1,12 +1,15 @@
 from elastic import Elastic
-ES_HOST = "http://localhost"
-ES_PORT = 9200
+ES= "http://localhost:9200"
 
-client = Elastic(ES_HOST, ES_PORT)
+client = Elastic(ES)
 indices = client.list_indices().keys()
 
 for index_name in indices:
+    if index_name.startswith("demo-empty"):
+        continue
+
     print(f"Removing index {index_name}")
+
     if index_name.startswith("indra"):
         client.set_readonly(index_name, False)
     client.delete_index(index_name)
